Refactored Salesforce.py to separate selenium, api, and performance keywords I also added BaseLibrary.py so that we have a single place where we define the helper properties (self.cumulusci, self.salesforce, etc).
This commit adds no keywords, it just reorganizes the code.
