+++ cumulusci/robotframework/Salesforce.robot
@@ -21,6 +21,7 @@ Library        OperatingSystem
 Library        String
 Library        XML
 Library        SeleniumLibrary  implicit_wait=${IMPLICIT_WAIT}  timeout=${TIMEOUT}
+Library        cumulusci.robotframework.SalesforceAPI
 Library        cumulusci.robotframework.CumulusCI  ${ORG}
 Library        cumulusci.robotframework.Salesforce  debug=${DEBUG}
